# minimize save/restore allocations